♻️ Refactor : Emoji 컴포넌트 공용 컴포넌트화
