+++ src/app/detail/panel/FeelingsCalendar.css
@@ -127,3 +127,10 @@ abbr[title] {
   display: inline-block;
   height: "0.0625rem";
 }
+
+.date img {
+  width: "100%";
+  height: "100%";
+  max-width: "2.875rem";
+  max-height: "2rem";
+}
